KeyObserver class
