+++ web/modules/dom/src/index.ts
@@ -3,6 +3,7 @@ export * from "./base-tooltip-button"
 export * from "./custom-elements"
 export * from "./focus"
 export * from "./gesture"
+export * from "./held"
 export * from "./hover"
 export * from "./key-handling"
 export * from "./observe"
